# docs/openapi.yaml
openapi: 3.0.3
info:
  title: NeuroWealth API
  version: 1.0.0
  description: |
    REST API for the NeuroWealth platform — AI-assisted portfolio management
    backed by Stellar smart contracts.

    **Versioning:** All endpoints are served under an explicit version prefix
    (`/api/v1/*`). The legacy unversioned paths (`/api/*`) remain available as
    deprecated aliases and return `Deprecation`/`Sunset` headers. Every response
    includes an `X-API-Version` header. See `docs/api-versioning.md`.

    **Breaking-change policy:** This API follows semantic versioning. Breaking
    changes (removed fields, changed response shapes, new required parameters)
    increment the major version and are announced at least two weeks before release.
servers:
  - url: http://localhost:{port}
    description: Local development
    variables:
      port:
        default: '3000'
  - url: https://api.neurowealth.app
    description: Production
...
security: []
paths:
...
  /api/v1/portfolio/{userId}:
    get:
      tags: [portfolio]
      operationId: getPortfolio
      summary: Get portfolio positions
      description: |
        Returns the user's portfolio including all active positions,
        total balance, and total earnings. The authenticated user can
        only access their own portfolio (enforceUserAccess).
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID v4)
      responses:
        '200':
          description: Portfolio data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioResponse'
              example:
                userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
                totalBalance: 12450.75
                totalEarnings: 523.40
                activePositions: 3
                positions:
                  - id: pos-001
                    protocolName: Aave
                    assetSymbol: USDC
                    currentValue: 10000.00
                    yieldEarned: 423.15
                    status: ACTIVE
                  - id: pos-002
                    protocolName: Compound
                    assetSymbol: XLM
                    currentValue: 2000.00
                    yieldEarned: 85.50
                    status: ACTIVE
                  - id: pos-003
                    protocolName: unassigned
                    assetSymbol: USDC
                    currentValue: 450.75
                    yieldEarned: 14.75
                    status: ACTIVE
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/transactions/{userId}:
    get:
      tags: [transactions]
      operationId: listTransactions
      summary: List user transactions
      description: Returns a paginated list of transactions for the given user.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
          description: User ID (UUID v4)
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-indexed)
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
          description: Items per page
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                type: object
                properties:
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
              example:
                page: 1
                limit: 5
                total: 42
                transactions:
                  - id: tx-001
                    txHash: a1b2c3d4e5f6...
                    type: DEPOSIT
                    status: CONFIRMED
                    amount: 1000.00
                    assetSymbol: USDC
                    protocolName: Aave
                    createdAt: '2026-06-28T10:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/deposit:
    post:
      tags: [deposit]
      operationId: initiateDeposit
      summary: Initiate an on-chain deposit
      description: |
        Submits a deposit transaction to the active protocol on behalf of the
        authenticated user. The user must match the userId in the request body.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, assetSymbol]
              properties:
                userId:
                  type: string
                  format: uuid
                  description: User ID (UUID v4)
                amount:
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                  description: Deposit amount
                assetSymbol:
                  type: string
                  minLength: 1
                  description: Asset symbol (e.g. USDC, XLM)
                protocolName:
                  type: string
                  description: Target protocol (optional, uses active protocol if omitted)
                memo:
                  type: string
                  maxLength: 280
                  description: Optional memo
            example:
              userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              amount: 1000.00
              assetSymbol: USDC
              protocolName: Aave
      responses:
        '201':
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
              example:
                txHash: a1b2c3d4e5f6...
                status: CONFIRMED
                transaction:
                  id: tx-001
                  txHash: a1b2c3d4e5f6...
                  status: CONFIRMED
                  amount: 1000.00
                  assetSymbol: USDC
                  protocolName: Aave
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Duplicate transaction hash
          $ref: '#/components/responses/Conflict'

  # ── Recurring Deposit ─────────────────────────────────────────────────────
  /api/v1/deposit/recurring:
    post:
      tags: [deposit]
      operationId: createRecurringDeposit
      summary: Create a recurring deposit plan
      description: |
        Creates a scheduled recurring deposit. Requires explicit confirmation.
        The deposit will execute automatically on the specified cadence.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, assetSymbol, cadence, confirmed]
              properties:
                userId:
                  type: string
                  format: uuid
                amount:
                  type: number
                  exclusiveMinimum: true
                assetSymbol:
                  type: string
                  minLength: 1
                cadence:
                  type: string
                  enum: [WEEKLY, BIWEEKLY, MONTHLY]
                confirmed:
                  type: boolean
                  enum: [true]
                  description: Must be true to confirm the recurring deposit
            example:
              userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              amount: 50
              assetSymbol: USDC
              cadence: WEEKLY
              confirmed: true
      responses:
        '201':
          description: Recurring deposit plan created
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/deposit/recurring/{id}:
    patch:
      tags: [deposit]
      operationId: updateRecurringDeposit
      summary: Update a recurring deposit plan
      description: Pause, resume, or update the amount/cadence of a plan.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  exclusiveMinimum: true
                cadence:
                  type: string
                  enum: [WEEKLY, BIWEEKLY, MONTHLY]
                status:
                  type: string
                  enum: [ACTIVE, PAUSED, CANCELLED]
      responses:
        '200':
          description: Plan updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [deposit]
      operationId: cancelRecurringDeposit
      summary: Cancel a recurring deposit plan
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Plan cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  plan:
                    $ref: '#/components/schemas/RecurringDepositPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Withdraw ───────────────────────────────────────────────────────────────
  /api/v1/withdraw:
    post:
      tags: [withdraw]
      operationId: initiateWithdrawal
      summary: Initiate an on-chain withdrawal
      description: |
        Submits a withdrawal transaction from the active protocol on behalf of
        the authenticated user. The user must match the userId in the request body.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, amount, assetSymbol]
              properties:
                userId:
                  type: string
                  format: uuid
                amount:
                  type: number
                  minimum: 0
                  exclusiveMinimum: true
                assetSymbol:
                  type: string
                  minLength: 1
                protocolName:
                  type: string
                memo:
                  type: string
                  maxLength: 280
            example:
              userId: a1b2c3d4-e5f6-7890-abcd-ef1234567890
              amount: 500.00
              assetSymbol: USDC
      responses:
        '201':
          description: Withdrawal initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionResponse'
              example:
                txHash: f6e5d4c3b2a1...
                status: CONFIRMED
                transaction:
                  id: tx-002
                  txHash: f6e5d4c3b2a1...
                  status: CONFIRMED
                  amount: 500.00
                  assetSymbol: USDC
                  protocolName: Aave
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'

  # ── Vault ──────────────────────────────────────────────────────────────────
...
  /api/v1/fiat/webhook/{provider}:
    post:
      tags: [fiat]
      operationId: fiatProviderWebhook
      summary: Fiat provider webhook callback
      description: >
        Signed callback delivered by the fiat provider on order state changes.
        Authenticity is verified from the provider's HMAC signature over the
        raw request body — no JWT is required. The endpoint is idempotent and
        always ACKs a well-formed, verified delivery. A provider "completed"
        signal advances the order to PROCESSING; settlement to SETTLED requires
        independent on-chain confirmation.
      parameters:
        - in: path
          name: provider
          required: true
          schema:
            type: string
          description: Provider key (e.g. `moonpay`)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Provider-specific payload (opaque; verified by signature)
      responses:
        '200':
          description: Delivery accepted (idempotent)
        '400':
          description: Malformed payload
        '401':
          description: Invalid or missing signature
        '404':
          description: Unknown provider
        '500':
          description: Processing error (provider should retry)

  # ── Referrals ────────────────────────────────────────────────────────────────
...
  /api/v1/referrals/{userId}:
    get:
      tags: [referrals]
      operationId: getReferrals
      summary: List referrals attributed to a user
      description: >
        Lists the referrals attributed to the user's code, newest first.
        Owner-scoped — a caller may only read their own referrals.
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: The owning user's id (must match the authenticated caller)
      responses:
        '200':
          description: Referrals for the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferralListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Caller may only read their own referrals
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token issued by `POST /api/v1/auth/verify`.
        Include as `Authorization: Bearer <token>`.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: API-Key
      description: >
        Admin API key. Can be provided either as
        `Authorization: Bearer <token>` or as the legacy
        `X-Admin-Token` header. Issued via `POST /api/v1/admin/keys`.
    InternalToken:
      type: apiKey
      in: header
      name: X-Internal-Token
      description: >
        Service-to-service token for internal endpoints (e.g., agent status,
        Prometheus metrics). Alternative auth methods include IP allowlisting
        or admin Bearer token.
  schemas:
...
    PortfolioResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        totalBalance:
          type: number
        totalEarnings:
          type: number
        activePositions:
          type: integer
        positions:
          type: array
          items:
            $ref: '#/components/schemas/Position'
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        txHash:
          type: string
        type:
          type: string
          enum: [DEPOSIT, WITHDRAWAL, SWAP, VAULT_DEPOSIT, VAULT_WITHDRAW]
        status:
          type: string
          enum: [PENDING, CONFIRMED, FAILED]
        amount:
          type: number
        assetSymbol:
          type: string
        protocolName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Not found
...